requirement.txt
pandas
numpy
pymysql
scikit-learn
python-dotenv
mysql-connector-python
nltk
Flask
dill
# -e .